Extract common logics of clients as abstract classes.
